+++ src/Cdn/Cdn/ChangeLog.md
@@ -20,6 +20,7 @@
 
 ## Upcoming Release
 * Added ChangeSafety Support
+* Added post-migration endpoint cutover guidance for Azure Front Door (classic) and Azure CDN from Microsoft (classic) migrations.
 
 ## Version 6.1.0
 * Upgraded API version to 2026-04-01-preview.